apps/api/pyproject.toml
[project]
name = "shadowtwin-api"
version = "0.1.0"
description = "Django API for ShadowTwin"
requires-python = ">=3.12"
dependencies = [
  "django>=5.1",
  "djangorestframework>=3.15",
  "django-environ>=0.11",
  "django-cors-headers>=4.6",
  "psycopg[binary]>=3.2",
  "PyJWT>=2.10",
  "redis>=5.2",
  "celery>=5.4",
  "sentry-sdk>=2.19"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
